+++ src/app/display_drivers.cpp
@@ -12,6 +12,8 @@
 #include "drivers/arduino_gfx_driver.cpp"
 #elif DISPLAY_DRIVER == DISPLAY_DRIVER_ARDUINO_GFX_ST77916
 #include "drivers/arduino_gfx_st77916_driver.cpp"
+#elif DISPLAY_DRIVER == DISPLAY_DRIVER_ARDUINO_GFX_NV3041A
+#include "drivers/arduino_gfx_nv3041a_driver.cpp"
 #elif DISPLAY_DRIVER == DISPLAY_DRIVER_ST7701_RGB
 #include "drivers/st7701_rgb_driver.cpp"
 #elif DISPLAY_DRIVER == DISPLAY_DRIVER_ST7703_DSI