+++ src/app/display_manager.cpp
@@ -21,6 +21,8 @@
 #include "drivers/arduino_gfx_driver.h"
 #elif DISPLAY_DRIVER == DISPLAY_DRIVER_ARDUINO_GFX_ST77916
 #include "drivers/arduino_gfx_st77916_driver.h"
+#elif DISPLAY_DRIVER == DISPLAY_DRIVER_ARDUINO_GFX_NV3041A
+#include "drivers/arduino_gfx_nv3041a_driver.h"
 #elif DISPLAY_DRIVER == DISPLAY_DRIVER_ST7701_RGB
 #include "drivers/st7701_rgb_driver.h"
 #elif DISPLAY_DRIVER == DISPLAY_DRIVER_ST7703_DSI
@@ -67,6 +69,8 @@ DisplayManager::DisplayManager(DeviceConfig* cfg)
 		driver = new Arduino_GFX_Driver();
 		#elif DISPLAY_DRIVER == DISPLAY_DRIVER_ARDUINO_GFX_ST77916
 		driver = new Arduino_GFX_ST77916_Driver();
+		#elif DISPLAY_DRIVER == DISPLAY_DRIVER_ARDUINO_GFX_NV3041A
+		driver = new Arduino_GFX_NV3041A_Driver();
 		#elif DISPLAY_DRIVER == DISPLAY_DRIVER_ST7701_RGB
 		driver = new ST7701_RGB_Driver();
 		#elif DISPLAY_DRIVER == DISPLAY_DRIVER_ST7703_DSI